+++ src/lib/debug_info/debug_info_view.js
@@ -54,6 +54,8 @@ ${translate('debugInfo.details.markChanges')}:      ${page.markChanges}
 ${translate('debugInfo.details.doPost')}:           ${page.doPost}
 ${translate('debugInfo.details.postParams')}:       ${page.postParams}
 ${translate('debugInfo.details.sendCredentials')}:  ${page.sendCredentials}
+${translate('debugInfo.details.waitForNetworkIdle')}: ${page.waitForNetworkIdle}
+${translate('debugInfo.details.waitForNetworkIdleTimeoutMs')}: ${page.waitForNetworkIdleTimeoutMs}
 ${translate('debugInfo.details.state')}:            ${page.state}
 ${translate('debugInfo.details.lastAutoscanTime')}: ${lastAutoscanTime}
 ${translate('debugInfo.details.oldScanTime')}:      ${oldScanTime}